requirements.txt
llama_cpp_python ~= 0.2.53
blessed ~= 1.20.0
prompt-toolkit ~= 3.0.43